bandwidth_limit: standarize in size_t
RH-Author: Juan Quintela <quintela@redhat.com>
Message-id: <1420649114-17435-11-git-send-email-quintela@redhat.com>
Patchwork-id: 63167
O-Subject: [PATCH qemu-kvm RHEL6.7 10/10] bandwidth_limit: standarize in size_t
Bugzilla: 970103
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Dr. David Alan Gilbert (git) <dgilbert@redhat.com>

Right now we:

- read from migrate_set_speed() as a double
- there we compare it with UINT32_MAX
- store it in a uint32_t
- then we call *_start_outgoing_migration() with that uint32_t
  but that functions take an int64_t
- inside that function we store that int64_t value into an int64_t
- we call migrate_fd_connect after several hoops
- we call there qemu_fopen_ops_buffered that expects a size_t
- we store there as xfer_limit, that is a size_t

To make things ever more interesting, the value is called:
- max_throttle (uint32_t)
- bandwdith_limit (int64_t)
- bytes_per_sec (size_t): until here, everything is bytes_per_sec
- xfer_limit (size_t): this one is bytes per 100ms

---
 migration-exec.c |  2 +-
 migration-fd.c   |  2 +-
 migration-tcp.c  |  2 +-
 migration-unix.c |  2 +-
 migration.c      |  4 ++--
 migration.h      | 10 +++++-----
 6 files changed, 11 insertions(+), 11 deletions(-)

+++ migration.c
@@ -40,7 +40,7 @@ static int64_t start, stop;
 #endif
 
 /* Migration speed throttling */
-static uint32_t max_throttle = (32 << 20);
+static size_t max_throttle = (32 << 20);
 
 static MigrationState *current_migration;
 
@@ -191,7 +191,7 @@ int do_migrate_set_speed(Monitor *mon, const QDict *qdict, QObject **ret_data)
     FdMigrationState *s;
 
     d = qdict_get_double(qdict, "value");
-    d = MAX(0, MIN(UINT32_MAX, d));
+    d = MAX(0, MIN(SIZE_MAX, d));
     max_throttle = d;
 
     s = migrate_to_fms(current_migration);
